Expose `Dependency`'s `kind` and `req` fields
`kind` has the specialised `DependencyKind`; `req` is just a string -
see https://github.com/steveklabnik/semver/issues/119
